refactor flask service
