+++ src/annotation/jobs/one_off_jobs/annotation_wf2_construct_all.py
@@ -1,6 +1,5 @@
 import datetime
 import hashlib
-
 import boto3
 from tqdm import tqdm
