+++ src/comms/sqs_comms.py
@@ -1,6 +1,5 @@
 import boto3
 import logging
-
 import time
