--- src/experiments/framework/sparse.py
@@ -1,8 +1,0 @@
-import scipy.sparse
-
-
-def sp_len(obj):
-    if scipy.sparse.issparse(obj):
-        return obj.shape[0]
-    else:
-        return len(obj)
